% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indice.R
\name{indice}
\alias{indice}
\title{indice}
\usage{
indice(spatial = NULL, date = NULL, s = NULL, d = NULL)
}
\arguments{
\item{spatial}{S4 spatial object}

\item{date}{object date from timeManip}

\item{s}{sub S4 spatial object}

\item{d}{sub object date from timeManip}
}
\description{
indice
}
\examples{
\dontrun{
indice()
}
}
\keyword{structify}

