% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.R
\name{structify_get}
\alias{structify_get}
\title{structify_get}
\usage{
structify_get(filename, vname, ...)
}
\arguments{
\item{filename}{filename}

\item{vname}{variable name "date", "spatial", "dataset",...}

\item{...}{other options according to the variable name}
}
\description{
structify_get
}
\examples{
\dontrun{
structify_get()
}
}
\keyword{structify}

