% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/put.R
\name{structify_put}
\alias{structify_put}
\title{structify_put}
\usage{
structify_put(filename, vname, obj, ..., NEW = FALSE)
}
\arguments{
\item{filename}{filename}

\item{vname}{variable name "date", "spatial", "dataset",...}

\item{obj}{object to be put}

\item{...}{other options according to the variable name}

\item{NEW}{boolean if new object}
}
\description{
structify_put
}
\examples{
\dontrun{
structify_get()
}
}
\keyword{structify}

