% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{structify_summary}
\alias{structify_summary}
\title{structify_summary}
\usage{
structify_summary(filename)
}
\arguments{
\item{filename}{filename}
}
\description{
structify_summary
}
\examples{
\dontrun{
structify_summary()
}
}
\keyword{structify}

